FunCineForge/requirements.txt
beartype==0.14.1
curl_cffi
dashscope
einops==0.8.1
face_alignment==1.4.1
fastcluster==1.2.6
funasr==1.2.9
gradio
g4f
hdbscan==0.8.40
huggingface-hub==0.34.3
kaldiio==2.18.1
librosa==0.10.1
matplotlib==3.7.5
ml_collections==0.1.1
modelscope==1.28.1
moviepy==1.0.3
numba==0.58.1
numpy==1.23.5
onnxruntime==1.19.2
openai
openai_whisper==20250625
opencc_python_reimplemented==0.1.7
opencv_python==4.12.0.88
packaging
pandas==2.0.3
pyannote.audio
pypinyin==0.44.0
python_speech_features==0.6
PyYAML==6.0.2
Requests
rotary_embedding_torch==0.8.9
scikit_learn==1.3.2
>scipy=1.10.0
soundfile==0.13.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
umap_learn==0.5.7
